requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
motor==3.3.2
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.10.6
pydantic-settings==2.0.0
websockets==11.0.3
python-dotenv==1.0.0
httpx==0.25.0
aiofiles==23.2.1
email-validator==2.1.1
slowapi==0.1.9
structlog==23.2.0
reportlab==4.0.7
authlib==1.2.1
httpx-oauth==0.13.1
itsdangerous==2.1.2
gunicorn==21.2.0
